+++ test/humanizer.js
@@ -72,7 +72,7 @@ describe('humanizer', function () {
     }), '1!!234 seconds')
   })
 
-  it('can round', function () {
+  it('can do simple rounding', function () {
     var h = humanizer({ round: true })
 
     assert.equal(h(0), '0 seconds')
@@ -84,6 +84,23 @@ describe('humanizer', function () {
     assert.equal(h(1500), '2 seconds')
     assert.equal(h(121499), '2 minutes, 1 second')
     assert.equal(h(121500), '2 minutes, 2 seconds')
+  })
+
+  it('can do rounding with the "units" option', function () {
+    var h = humanizer({ round: true })
+
+    assert.equal(h(86364000, { units: ['y', 'mo', 'w', 'd', 'h'] }), '1 day')
+    assert.equal(h(1209564000, { units: ['y', 'mo', 'w', 'd', 'h'] }), '2 weeks')
+    assert.equal(h(3692131200000, { units: ['y', 'mo'] }), '117 years')
+    assert.equal(h(3692131200001, { units: ['y', 'mo', 'w', 'd', 'h', 'm'] }), '116 years, 11 months, 4 weeks, 1 day, 4 hours, 30 minutes')
+  })
+
+  it('can do rounding with the "largest" option', function () {
+    var h = humanizer({ round: true })
+
+    assert.equal(h(3692131200000, { largest: 1 }), '117 years')
+    assert.equal(h(3692131200000, { largest: 2 }), '117 years')
+    assert.equal(h(3692131200001, { largest: 100 }), '116 years, 11 months, 4 weeks, 1 day, 4 hours, 30 minutes')
   })
 
   it('can ask for the largest units', function () {